backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
aiosqlite==0.22.1
    # via -r requirements.in
alembic==1.17.2
    # via -r requirements.in
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
bcrypt==5.0.0
    # via passlib
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
cffi==2.0.0
    # via cryptography
click==8.3.1
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
cryptography==46.0.3
    # via python-jose
ecdsa==0.19.1
    # via python-jose
fastapi==0.128.0
    # via -r requirements.in
greenlet==3.3.0
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via -r requirements.in
idna==3.11
    # via
    #   anyio
    #   httpx
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via mako
passlib[bcrypt]==1.7.4
    # via -r requirements.in
pillow==12.1.0
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycparser==2.23
    # via cffi
pydantic==2.12.5
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via -r requirements.in
python-dotenv==1.2.1
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
python-jose[cryptography]==3.5.0
    # via -r requirements.in
python-multipart==0.0.21
    # via -r requirements.in
python-slugify==8.0.4
    # via -r requirements.in
pyyaml==6.0.3
    # via uvicorn
rsa==4.9.1
    # via python-jose
six==1.17.0
    # via ecdsa
sqlalchemy==2.0.45
    # via
    #   -r requirements.in
    #   alembic
starlette==0.50.0
    # via fastapi
text-unidecode==1.3
    # via python-slugify
typing-extensions==4.15.0
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
uvicorn[standard]==0.40.0
    # via -r requirements.in
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via uvicorn
